Comment out high-level random-access event retrieval
Formats (such as S3M) can store pattern data in a format optimized for
sequential access, making calls that require random access to events
hard to implement efficiently. We plan to change that for a call that
retrieves the entire pattern data at once to a buffer, leaving for the
caller to format it for display.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>