[mixer] Always interpolate on 32-bit data
Change interpolators to work on data from a 32-bit buffer instead of
receiving data in 8-bit or 16-bit depending on sample size. Data is
added to the buffer as it is sampled.

This also allows us to remove the ugly extra bytes added around
sample data, and fixes interpolation at loop boundaries.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>